refactor: share location key helpers
